+++ src/Rendering/Renderers/PhotonMapVisualizer.h
@@ -8,7 +8,7 @@ public:
 	glm::vec3 GetPixelColor(const Ray & ray) override;
 	PhotonMapVisualizer(Scene & scene, const unsigned int PHOTONS_PER_LIGHT_SOURCE = 1000000,
 						const unsigned int MIN_PHOTONS_PER_NODE = 50,
-						const unsigned int MIN_DIMENSION_SIZE_OF_NODE = 0.2f,
+						const float MIN_DIMENSION_SIZE_OF_NODE = 0.2f,
 						const unsigned int MAX_PHOTON_DEPTH = 3);
 private:
 	glm::vec3 TraceRay(const Ray & ray, const unsigned int DEPTH = 0);
